# doc/api.yaml
openapi: 3.0.3
info:
  title: WASA HW
  description: |-
    This OpenAPI document describes the WASA HW 
    OpenAPI specifications.
  version: "1"
...
paths:
...
  /users/{userName}:
    parameters: 
      - $ref : "#/components/parameters/username"
    get:
      tags: ["user profile"]
      summary: Get the user profile
      security: 
        - BearerAuth: []
      description: |-
        Fetch all the required informations
        for a user profile page, including
        followers, following, how many photos
        have been uploaded. If user is banned
        by userName contained in path then
        the action is forbidden.
      operationId: getUserProfile
      responses:
        '200':
          description: User fetch action successful
          content:
            application/json:
              schema: { $ref : "#/components/schemas/Profile" }
              example:
                following:
                  - user_id: "MichaelJackson"
                  - user_id: "IndianaJones"
                followers: 
                  - user_id: "Coez"
                  - user_id: "MaraFrascarelli"
                photos:
                  - photo_id: 1
                    author_id: "GabrieleLorenzini"
                    date: 2023-09-24T17:23:34Z
                  - photo_id: 2
                    author_id: "DavideFrattafalso"
                    date: 2023-09-24T17:33:34Z
        '401': { $ref : "#/components/responses/Unauthorized" }
        '403': { $ref : "#/components/responses/Forbidden" }
        '404': { $ref : "#/components/responses/NotFound" }
        '500': { $ref : "#/components/responses/InternalServerError" }
    put:
      tags: ["username"]
      summary: Sets the user name
      security: 
        - BearerAuth: []
      description: |-
        Change the actual user name to the one
        specified in the body and return a new
        identifier.
      operationId: setMyUserName
      requestBody:
        description: New user name
        content:
          application/json:
            schema: { $ref : "#/components/schemas/Username" }
            example: 
              name: Maria
        required: true
      responses:
        '200':
          description: Set user name action successful
          content:
            application/json:
              schema: { $ref : "#/components/schemas/Identifier" }
              example: 
                identifier: "abcdef012345"
        '400': { $ref : "#/components/responses/BadRequest" }
        '500': { $ref : "#/components/responses/InternalServerError" }
...
  /users/{userName}/photos:
    parameters: 
      - $ref : "#/components/parameters/username"
    get:
      tags: ["user photo"]
      summary: Get all photos of an user
      description: |-
        Get all the photos posted by the
        user specified in the path.
      operationId: getPhotos
      responses:
        '200':
          description: Get a list of photos
          content:
            application/json:
              schema: { $ref : "#/components/schemas/Photos" }
              example:
                photos:
                  - photo_id: 1
                    author_id: "Sfera99"
                    date: 2023-09-24T17:23:34Z
                  - photo_id: 2
                    author_id: "Gue98"
                    date: 2023-09-24T17:23:34Z
        '403': { $ref : "#/components/responses/Forbidden" }
        '500': { $ref : "#/components/responses/InternalServerError" }
    post:
      tags: ["user photo"]
      summary: Upload a photo
      security:
        - BearerAuth: [] 
      description: |-
        Uploads a photo, date and time are calculated
        by the server when the request is made.
        Only the users by the ID specified in the header can
        upload a photo.
      operationId: uploadPhoto
      requestBody:
        description: Photo details
        content:
          multipart/form-data:
            schema: { $ref : "#/components/schemas/PhotoBin" }
        required: true
      responses:
        '201': 
          description: Photo upload action successful
          content:
            application/json:
              schema: { $ref : "#/components/schemas/Photo" }
              example:
                photo_id: 1
                author_id: Roberto4Coni
                date: 2023-09-24T17:23:34Z
        '500': { $ref : "#/components/responses/InternalServerError" }
...
components:
  parameters:
    username:
      in: path
      name: userName
      required: true
      schema: { $ref: '#/components/schemas/Username/properties/name' }
      description: Session ID of a user
      example: Maria
...
  responses:
...
    BadRequest:
      description: Response with status code 400 (HTTP Bad Request)
    Unauthorized:
      description: Response with status code 401 (HTTP Unauthorized)
    Forbidden:
      description: Response with status code 403 (HTTP Forbidden)
    NotFound:
      description: Response with status code 404 (HTTP Not Found)
    InternalServerError:
      description: Response with status code 500 (HTTP Internal Server Error)
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
    Identifier:
      description: |-
        A schema that describes an
        identifier for an user
      type: object
      properties:
        identifier:
          description: Unique ID of an user
          type: string
          example: "abcdef012345"
    Username:
      description: |-
        A schema that describes the
        name of the user.
      type: object
      properties:
        name:
          description: Name of an user
          type: string
          minLength: 3
          maxLength: 16
          pattern: '^.*?$'
          example: 'Maria'
    Photos:
      description: |-
        A schema that describes
        a collection of photos
      type: object
      properties:
        photos:
          description: Array of photos 
          type: array
          minItems: 0
          maxItems: 100
          items:
            $ref : '#/components/schemas/Photo'
    PhotoBin:
      description: |-
        A schema that describes a
        photo object converted in binary
      type: object
      properties:
        file:
          description: Binary file
          type: string
          format: binary
          minLength: 0
          maxLength: 100
    Photo:
      description: |-
        A object that contains all the useful
        info of a photo object
      type: object
      properties: 
        photo_id:
          description: Photo unique ID
          type: integer
          example: 484
        author_id:
          $ref : '#/components/schemas/Identifier/properties/identifier'
        date:
          description: DateTime of the uploaded photo
          type: string
          format: date-time
          example: 2023-09-24T17:23:34Z
    Profile:
      description: Profile of an user
      type: object
      properties:
        following: { $ref: "#/components/schemas/Follow" }
        followers: { $ref: "#/components/schemas/Follow" }
        photos: { $ref: "#/components/schemas/Photos/properties/photos" }
    Follow:
      description: Array of user that are followers or following
      type: array
      minItems: 0
      maxItems: 999
      items: { $ref: "#/components/schemas/User" }
    User:
      description: Info about an user such as User ID
      type: object
      properties:
        user_id: { $ref: "#/components/schemas/Username/properties/name" }
...